+++ app/src/main/java/com/eatssu/android/view/review/MyReviewDialogActivity.kt
@@ -1,10 +1,18 @@
 package com.eatssu.android.view.review
 
+import android.content.DialogInterface
 import android.content.Intent
 import android.os.Bundle
 import android.util.Log
+import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.lifecycleScope
+import com.eatssu.android.data.service.ReviewService
 import com.eatssu.android.databinding.ActivityMyReviewDialogBinding
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 
 class MyReviewDialogActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMyReviewDialogBinding
@@ -27,6 +35,49 @@ class MyReviewDialogActivity : AppCompatActivity() {
             intent.putExtra("menu", menu)
             startActivity(intent)
             finish()
+        }
+
+        binding.btnReviewDelete.setOnClickListener {
+
+            AlertDialog.Builder(this).apply {
+                setTitle("리뷰 삭제")
+                setMessage("작성한 리뷰를 삭제하시겠습니까?")
+                setNegativeButton("취소", DialogInterface.OnClickListener{ dialog, which ->
+                    Toast.makeText(this@MyReviewDialogActivity, "리뷰 삭제가 취소되었습니다", Toast.LENGTH_SHORT).show()
+                })
+                setPositiveButton("삭제", DialogInterface.OnClickListener { dialog, which ->
+                    val service = RetrofitImpl.retrofit.create(ReviewService::class.java)
+                    lifecycleScope.launch {
+                        try {
+                            val response = withContext(Dispatchers.IO) {
+                                service.delReview(reviewId).execute()
+                            }
+                            if (response.isSuccessful) {
+                                // 정상적으로 통신이 성공한 경우
+                                Log.d("post", "onResponse 성공: " + response.body().toString())
+                                Toast.makeText(
+                                    this@MyReviewDialogActivity, "삭제하기가 완료되었습니다.", Toast.LENGTH_SHORT
+                                ).show()
+                                finish()
+                            } else {
+                                // 통신이 실패한 경우(응답코드 3xx, 4xx 등)
+                                Log.d(
+                                    "post",
+                                    "onResponse 실패 write" + response.code()
+                                )
+                            }
+                        } catch (e: Exception) {
+                            // 통신 중 예외가 발생한 경우
+                            Log.d("post", "통신 실패: ${e.message}")
+                            Toast.makeText(
+                                this@MyReviewDialogActivity, "리뷰를 삭제하지 못했습니다.", Toast.LENGTH_SHORT
+                            ).show()
+                        }
+                    }
+                })
+                create()
+                show()
+            }
         }
     }
 }
